refactor: rename stac_model file to stac_widget
