+++ packages/stac_models/lib/core/stac_widget.dart
