+++ packages/stac_models/lib/painting/stac_edge_insets/stac_edge_insets.dart
@@ -1,5 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
-import 'package:stac_models/core/stac_model.dart';
+import 'package:stac_models/core/stac_widget.dart';
 
 part 'stac_edge_insets.g.dart';
