+++ packages/stac_models/lib/painting/stac_system_ui_overlay_style/stac_system_ui_overlay_style.dart
@@ -1,5 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
-import 'package:stac_models/core/stac_model.dart';
+import 'package:stac_models/core/stac_widget.dart';
 
 part 'stac_system_ui_overlay_style.g.dart';
