+++ packages/stac_models/lib/painting/stac_text_span/stac_text_span.dart
@@ -1,5 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
-import 'package:stac_models/core/stac_model.dart';
+import 'package:stac_models/core/stac_widget.dart';
 import 'package:stac_models/painting/stac_text_style/stac_text_style.dart';
 
 part 'stac_text_span.g.dart';
