+++ packages/stac_models/lib/stac_models.dart
@@ -1,6 +1,6 @@
 library;
 
-export 'core/stac_model.dart';
+export 'core/stac_widget.dart';
 export 'painting/painting.dart';
 export 'types/types.dart';
 export 'widgets/widgets.dart';
