+++ packages/stac_models/lib/theme/stac_button_style/stac_button_style.dart
@@ -1,5 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
-import 'package:stac_models/core/stac_model.dart';
+import 'package:stac_models/core/stac_widget.dart';
 import 'package:stac_models/painting/stac_color/stac_colors.dart';
 import 'package:stac_models/painting/stac_edge_insets/stac_edge_insets.dart';
 import 'package:stac_models/painting/stac_text_style/stac_text_style.dart';
