+++ packages/stac_models/lib/types/stac_border/stac_border.dart
@@ -1,5 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
-import 'package:stac_models/core/stac_model.dart';
+import 'package:stac_models/core/stac_widget.dart';
 import 'package:stac_models/painting/stac_color/stac_colors.dart';
 import 'package:stac_models/types/stac_border_side/stac_border_side.dart';
