+++ packages/stac_models/lib/types/stac_border_radius/stac_border_radius.dart
@@ -1,5 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
-import 'package:stac_models/core/stac_model.dart';
+import 'package:stac_models/core/stac_widget.dart';
 
 part 'stac_border_radius.g.dart';
