+++ packages/stac_models/lib/types/stac_offset/stac_offset.dart
@@ -1,5 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
-import 'package:stac_models/core/stac_model.dart';
+import 'package:stac_models/core/stac_widget.dart';
 
 part 'stac_offset.g.dart';
