+++ packages/stac_models/lib/types/stac_rect/stac_rect.dart
@@ -1,6 +1,6 @@
 import 'package:json_annotation/json_annotation.dart';
 import 'package:stac_models/core/converters/double_converter.dart';
-import 'package:stac_models/core/stac_model.dart';
+import 'package:stac_models/core/stac_widget.dart';
 import 'package:stac_models/types/stac_offset/stac_offset.dart';
 
 part 'stac_rect.g.dart';
