+++ packages/stac_models/lib/types/stac_visual_density/stac_visual_density.dart
@@ -1,6 +1,6 @@
 import 'package:json_annotation/json_annotation.dart';
 import 'package:stac_models/core/converters/double_converter.dart';
-import 'package:stac_models/core/stac_model.dart';
+import 'package:stac_models/core/stac_widget.dart';
 
 part 'stac_visual_density.g.dart';
