+++ packages/stac_models/lib/widgets/expanded/stac_expanded.dart
@@ -1,5 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
-import 'package:stac_models/core/stac_model.dart';
+import 'package:stac_models/core/stac_widget.dart';
 import 'package:stac_models/core/converters/widget_converter.dart';
 
 part 'stac_expanded.g.dart';
