+++ packages/stac_models/lib/widgets/floating_action_button/stac_floating_action_button.dart
@@ -1,6 +1,6 @@
 import 'package:json_annotation/json_annotation.dart';
 import 'package:stac_models/core/converters/widget_converter.dart';
-import 'package:stac_models/core/stac_model.dart';
+import 'package:stac_models/core/stac_widget.dart';
 import 'package:stac_models/painting/stac_text_style/stac_text_style.dart';
 import 'package:stac_models/types/stac_floating_action_button_type.dart';
