+++ packages/stac_models/lib/widgets/fractionally_sized_box/stac_fractionally_sized_box.dart
@@ -1,5 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
-import 'package:stac_models/core/stac_model.dart';
+import 'package:stac_models/core/stac_widget.dart';
 import 'package:stac_models/core/converters/double_converter.dart';
 import 'package:stac_models/core/converters/widget_converter.dart';
 import 'package:stac_models/types/stac_alignment.dart';
