+++ packages/stac_models/lib/widgets/positioned/stac_positioned.dart
@@ -1,5 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
-import 'package:stac_models/core/stac_model.dart';
+import 'package:stac_models/core/stac_widget.dart';
 import 'package:stac_models/core/converters/double_converter.dart';
 import 'package:stac_models/core/converters/widget_converter.dart';
