+++ packages/stac_models/lib/widgets/stack/stac_stack.dart
@@ -1,5 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
-import 'package:stac_models/core/stac_model.dart';
+import 'package:stac_models/core/stac_widget.dart';
 import 'package:stac_models/types/stac_alignment.dart';
 import 'package:stac_models/types/stac_text_types.dart';
 import 'package:stac_models/types/stac_stack_fit.dart';
