+++ packages/stac_models/lib/widgets/text/stac_text.dart
@@ -1,5 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
-import 'package:stac_models/core/stac_model.dart';
+import 'package:stac_models/core/stac_widget.dart';
 import 'package:stac_models/painting/stac_color/stac_colors.dart';
 import 'package:stac_models/painting/stac_text_span/stac_text_span.dart';
 import 'package:stac_models/painting/stac_text_style/stac_text_style.dart';
